move dashboard to main blueprint + change testing/dev mail provider
