+++ app/__init__.py
@@ -21,7 +21,6 @@ login_manager = LoginManager()
 login_manager.session_protection = 'strong'
 login_manager.login_view = 'account.login'
 
-
 def create_app(config_name="dev"):
     """ Create the application """
     app = Flask(
