+++ app/main/__init__.py
@@ -4,4 +4,4 @@ from flask import Blueprint
 main = Blueprint('main', __name__)
 
 from . import errors, filters
-from .views import tokens, corpus, index
+from .views import tokens, corpus, index, dashboard
