+++ app/templates/admin/new_user.html
@@ -8,7 +8,7 @@
 {% block content %}
     <div class="container">
         <div>
-            <a class="compact button" href="{{ url_for('admin.dashboard') }}">
+            <a class="compact button" href="{{ url_for('main.dashboard') }}">
                 <i class="caret left icon"></i>
                 Back to dashboard
             </a>
