+++ app/templates/admin/registered_users.html
@@ -4,7 +4,7 @@
 {% block content %}
     <div class="ui stackable grid container">
         <div class="sixteen wide tablet twelve wide computer centered column">
-            <a class="ui basic compact button" href="{{ url_for('admin.dashboard') }}">
+            <a class="ui basic compact button" href="{{ url_for('main.dashboard') }}">
                 <i class="caret left icon"></i>
                 Back to dashboard
             </a>
