+++ app/templates/macros/tokens_macros.html
@@ -162,6 +162,7 @@
     <table class="table table-bordered">
         <thead>
             <tr>
+                <th>User</th>
                 <th>Edit</th>
                 <th>Form</th>
                 <th>Context</th>
@@ -174,6 +175,7 @@
         <tbody>
             {% for token in tokens.items %}
                 <tr>
+                    <td class="small">{{ token.user.first_name[0] }}.{{ token.user.last_name }}</td>
                     <td class="small">{{token.created_on}}</td>
                     <td>{{token.form}}</td>
                     <td class="small">{{token_context(token.word_token)}}</td>
