+++ app/templates/main/dashboard.html
@@ -1,0 +1,24 @@
+{% extends 'layouts/base.html' %}
+{% import 'macros/dashboard_macros.html' as dashboard %}
+
+
+{% block content %}
+    <div class="container">
+        {% if current_user.is_admin() %}
+            <div class="list-group">
+                <a href="#" class="list-group-item active bg-dark" style="border-color: darkgrey;">
+                    <h2>Admin Dashboard</h2>
+                </a>
+                        {{ dashboard.dashboard_option('Registered Users', 'admin.registered_users',
+                                        description='View and manage user accounts', icon='fa fa-users') }}
+                        {{ dashboard.dashboard_option('Add New User', 'admin.new_user',
+                                        description='Create a new user account', icon='fa fa-user-plus') }}
+                        {{ dashboard.dashboard_option('Invite New User', 'admin.invite_user',
+                                        description='Invites a new user to create their own account',
+                                        icon='fa fa-user-plus') }}
+
+            </div>
+        {% endif %}
+        {{ dashboard.corpora_users(current_user, corpora) }}
+    </div>
+{% endblock %}
