+++ app/templates/main/dashboard_manage_corpus_users.html
@@ -4,11 +4,11 @@
 {% block content %}
     <div class="container" id="main-container">
         <div class="mt-2" >
-            <a href="{{ url_for(".dashboard") }}">Back to dashboard</a>
+            <a href="{{ url_for("main.dashboard") }}">Back to dashboard</a>
         </div>
 
         <form method="POST" id="accesses-form"
-              action="{{ url_for("account.manage_corpus_users", corpus_id=corpus.id) }}">
+              action="{{ url_for("main.manage_corpus_users", corpus_id=corpus.id) }}">
             <div class="mt-3">
                 <h2><i class="fa fa-users mr-2"></i> View and manage corpus users</h2>
                 <p class="form-text">Add or remove corpus access for users</p>
